Added some missing #ifdef OOMPH_HAS_PARALLEL so that the code still compiles in serial.
Updated a few convergence tolerances so that the test pass on the work desktop

Found a tricky little floating point problem with the Z2 error estimator for
hp adaptive elements. The problem is that the integration scheme needs to be
of high enough order that the mass matrix of the recovery shape functions
is integrated exactly. If not, the mass matrix is nearly singular (numerically
 speaking) which causes problems in the LU decomposition. The integration 
scheme used to be taken from the elements, which was not always of high enough
order. Moreover, an element author should not need to know that the 
integration scheme is also going to be used by the error estimator. 
 In the new implementation, the integration scheme is created by the error 
estimator (although the element type --- quad or triangle) must still be 
found from the element. The resulting mass matrices are still often near 
singular for high order recovery shape functions, and the
more complete fix is to use a better set of (orthogonal polynomial) shape
functions for the polynomial space. For now, I haven't done this, but it 
should be done soon, particularly given that the error estimator is not
really appropriate for the high-order elements.

The validation data for the hp-adaptive tests has been regenerated.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@268 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/generic/hp_refineable_elements.cc
@@ -677,7 +677,7 @@ local_coordinate_of_node(const unsigned& n, Vector<double>& s)
   s.resize(2);
   unsigned Nnode_1d = this->nnode_1d();
   unsigned n0 = n%Nnode_1d;
-  unsigned n1 = unsigned(double(n)/double(Nnode_1d));
+  unsigned n1 = n/Nnode_1d; 
   
   switch(Nnode_1d)
    {
